test: refactor - dynamic field / image field
